requirements-dev.txt
-r requirements.txt
django-stubs==6.0.2
djangorestframework-stubs==3.16.9
mypy==1.20.1
pre-commit
pytest
pytest-django
pytest-sugar
ruff